fix: fix bugs related plan cached; add a new primitive
